chore: migrate the view UserPrederencesView to ts
